Improve handling of back navigation.
I'm not sure whether the original implementation was chosen to mimic
Prestel's historic behaviour or suchlike, but it felt rather unintuitive
to me when browsing the service.  Consider the following sequence from a
fresh start:

`0a -> 1a -> 11a -> [back] -> 12a -> [back]`

The first `[back]` takes the user to frame `1a`, as expected.  However,
the second one currently takes the user back to frame `0a` instead,
which came as somewhat of a surprise when I noticed what was happening!
My expectation would be for it to take user back to frame `1a` as well,
which is the behaviour introduced here.

The implementation is to change the back nav stack to always leave the
current page on the top of the stack, and discard it when navigating
backwards.  Then we can leave the new current page on top of the stack
and peek at it as many times as we need to, only discarding it when the
user navigates back once more from that entry.

It might be possible to simplify this somewhat, since currently there's
a `HistoryPage` flag which seems solely to be used to avoid pushing the
new current page onto the history stack once rendered - switching to a
pop instead of a peek might allow this special case to be removed?  I
need to stare at this some more to convince myself.

This also breaks the tests for back navigation, so I've fixed those up
by adding a new state to represent the new behaviour even though I don't
fully understand the old behaviour yet!